@@ -1,15 +1,15 @@
 module github.com/domonda/go-function/cmd/gen-func-wrappers
 
-go 1.17
+go 1.18
 
 require (
-	github.com/ungerik/go-astvisit v0.0.0-20220207183704-334a95c6784c
-	golang.org/x/tools v0.1.9
+	github.com/ungerik/go-astvisit v0.0.0-20220317123017-e5279ac0952d
+	golang.org/x/tools v0.1.10
 )
 
 require (
-	golang.org/x/mod v0.5.1 // indirect
-	golang.org/x/sys v0.0.0-20220204135822-1c1b9b1eba6a // indirect
+	golang.org/x/mod v0.6.0-dev.0.20220106191415-9b9b3d81d5e3 // indirect
+	golang.org/x/sys v0.0.0-20220317061510-51cd9980dadf // indirect
 	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 )
 
