@@ -1,6 +1,6 @@
 module github.com/domonda/go-function
 
-go 1.17
+go 1.18
 
 require (
 	github.com/fatih/color v1.13.0
@@ -11,5 +11,5 @@ require (
 require (
 	github.com/mattn/go-colorable v0.1.12 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
-	golang.org/x/sys v0.0.0-20220128215802-99c3d69c2c27 // indirect
+	golang.org/x/sys v0.0.0-20220317061510-51cd9980dadf // indirect
 )
