@@ -6,7 +6,7 @@ import (
 	"reflect"
 )
 
-func CallFunctionWithJSONArgs(ctx context.Context, f Wrapper, jsonObject []byte) (results []interface{}, err error) {
+func CallFunctionWithJSONArgs(ctx context.Context, f Wrapper, jsonObject []byte) (results []any, err error) {
 	args, err := unmarshalJSONFunctionArgs(f, jsonObject)
 	if err != nil {
 		return nil, err
@@ -14,13 +14,13 @@ func CallFunctionWithJSONArgs(ctx context.Context, f Wrapper, jsonObject []byte)
 	return f.Call(ctx, args)
 }
 
-func unmarshalJSONFunctionArgs(f Description, jsonObject []byte) (args []interface{}, err error) {
+func unmarshalJSONFunctionArgs(f Description, jsonObject []byte) (args []any, err error) {
 	argsJSON := make(map[string]json.RawMessage)
 	err = json.Unmarshal(jsonObject, &argsJSON)
 	if err != nil {
 		return nil, err
 	}
-	args = make([]interface{}, f.NumArgs())
+	args = make([]any, f.NumArgs())
 	argTypes := f.ArgTypes()
 	for i, argName := range f.ArgNames() {
 		argType := argTypes[i]
